% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_defaults-save.R
\name{ch_defaults_save}
\alias{ch_defaults_save}
\title{Saves the current defaults in a yaml file that is compatible with
the config package}
\usage{
ch_defaults_save(path = "chattr.yml", overwrite = FALSE, type = NULL)
}
\arguments{
\item{path}{Path to the file to save the configuration to}

\item{overwrite}{Indicates to replace the file if it exists}

\item{type}{The type of UI to save the defaults for. It defaults to NULL which
will save whatever types had been used during the current R session}
}
\description{
Saves the current defaults in a yaml file that is compatible with
the config package
}
