% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/history.R
\name{ch_history}
\alias{ch_history}
\title{Displays the current session' chat history}
\usage{
ch_history(x = NULL)
}
\arguments{
\item{x}{An list object that contains chat history. Use this argument to
override the current history.}
}
\description{
Displays the current session' chat history
}
