% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_test.R
\name{ch_test}
\alias{ch_test}
\title{Confirms connectivity with model}
\usage{
ch_test(defaults = ch_defaults())
}
\arguments{
\item{defaults}{Defaults object, generally puled from \code{ch_defaults()}}
}
\description{
Confirms connectivity with model
}
