% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch_use.R
\name{ch_use_openai_gpt35}
\alias{ch_use_openai_gpt35}
\alias{ch_use_openai_davinci}
\alias{ch_use_llamagpt}
\title{Sets base defaults for specific LLM model}
\usage{
ch_use_openai_gpt35()

ch_use_openai_davinci()

ch_use_llamagpt()
}
\description{
Sets base defaults for specific LLM model
}
