% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chattr.R
\name{chattr}
\alias{chattr}
\title{Submits prompt to LLM}
\usage{
chattr(prompt = NULL, preview = FALSE, prompt_build = TRUE, stream = NULL)
}
\arguments{
\item{prompt}{Request to send to LLM. Defaults to NULL}

\item{preview}{Primarily used for debugging. It indicates if it should send
the prompt to the LLM (FALSE), or if it should print out the resulting
prompt (TRUE)}

\item{prompt_build}{Include the context and additional prompt as part of the
request}

\item{stream}{To output the response from the LLM as it happens, or wait until
the response is complete. Defaults to TRUE.}
}
\description{
Submits prompt to LLM
}
