% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chattr-use.R
\name{chattr_use}
\alias{chattr_use}
\title{Sets the LLM model to use in your session}
\usage{
chattr_use(model_label = NULL)
}
\arguments{
\item{model_label}{The label of the LLM model to use. Valid values are
'copilot', 'gpt4', 'gpt35', 'davinci', and 'llamagpt'.}
}
\description{
Sets the LLM model to use in your session
}
\details{
Use the 'CHATTR_MODEL' environment variable to set it for the
R session.
}
