% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tc_base_prompt}
\alias{tc_base_prompt}
\title{Recommended base prompt}
\format{
Character vectors
}
\usage{
tc_base_prompt
}
\description{
It is meant to be used within the tidychat_default() function
}
\examples{

tc_defaults(
  c("This is an example", "{tc_base_prompt}"),
  type = "chat"
)

}
\keyword{datasets}
