% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults-save.R
\name{tc_defaults_save}
\alias{tc_defaults_save}
\title{Saves the current defaults in a yaml file that is compatible with
the config package}
\usage{
tc_defaults_save(path = "tidychat.yml", overwrite = FALSE, type = NULL)
}
\arguments{
\item{path}{Path to the file to save the configuration to}

\item{overwrite}{Indicates to replace the file if it exists}

\item{type}{The type of UI to save the defaults for. It defaults to NULL which
will save whatever types had been used during the current R session}
}
\description{
Saves the current defaults in a yaml file that is compatible with
the config package
}
