% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{tc_test}
\alias{tc_test}
\title{Confirms connectivity with model}
\usage{
tc_test(defaults = tc_defaults())
}
\arguments{
\item{defaults}{Defaults object, generally puled from \code{tc_defaults()}}
}
\description{
Confirms connectivity with model
}
