% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidychat_base_prompt}
\alias{tidychat_base_prompt}
\title{Recommended base prompt}
\format{
Character vectors
}
\usage{
tidychat_base_prompt
}
\description{
It is meant to be used within the tidychat_default() function
}
\examples{

tidychat_defaults(
  c("This is an example", "{tidychat_base_prompt}"),
  type = "chat"
)

}
\keyword{datasets}
