% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychat.R
\name{tidychat_base_yaml}
\alias{tidychat_base_yaml}
\title{Copies the base version of defaults used for tidychat}
\usage{
tidychat_base_yaml(overwrite = FALSE)
}
\arguments{
\item{overwrite}{If there's an existing "config.yml", should it be replaced?}
}
\description{
Copies the base version of defaults used for tidychat
}
