% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{tidychat_defaults}
\alias{tidychat_defaults}
\title{Default arguments to use when making requests to the LLM}
\usage{
tidychat_defaults(
  prompt = NULL,
  include_data_files = NULL,
  include_data_frames = NULL,
  include_doc_contents = NULL,
  provider = NULL,
  model = NULL,
  system_msg = NULL,
  yaml_file = "config.yml",
  model_arguments = NULL,
  type = "notebook"
)
}
\arguments{
\item{prompt}{Request to send to LLM. Defaults to NULL}

\item{include_data_files}{Send the list of data files found in the working
directory? Defaults to NULL}

\item{include_data_frames}{Send the list of data.frames currently in the R
environment? Defaults to NULL}

\item{include_doc_contents}{Send the current code in the document}

\item{provider}{The name of the provider of the LLM. Today, only "openai" is
is available}

\item{model}{The name of the model to use, based on the provider}

\item{system_msg}{For GPT 3.5 or above, the system message to send as part of
the request}

\item{yaml_file}{The path to a valid \code{config} YAML file that contains the
defaults to use in a session}

\item{model_arguments}{Additional arguments to pass to the model as part of the
request, it requires a list. Examples of arguments: temperature, top_p,
max_tokens}

\item{type}{Entry point to interact with the model. Accepted values: 'notebook',
'chat'}
}
\description{
Default arguments to use when making requests to the LLM
}
\details{
The idea is that because we will use addin shortcut to execute the
request, all of the other arguments can be controlled via this function. By
default, it will try to load defaults from a \code{config} YAML file, if none are
found, then the defaults for GPT 3.5 will be used. The defaults can be modified
by calling this function, even after the interactive session has started.
}
