% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychat.R
\name{tidychat_prompt}
\alias{tidychat_prompt}
\alias{tidychat_debug_get}
\alias{tidychat_history}
\title{Debugging functions}
\usage{
tidychat_prompt(prompt = NULL)

tidychat_debug_get()

tidychat_history(raw = FALSE)
}
\arguments{
\item{prompt}{Request to send to LLM. Defaults to NULL}
}
\description{
Debugging functions
}
