% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{tidychat_submit}
\alias{tidychat_submit}
\title{Method to easily integrate to new LLM's}
\usage{
tidychat_submit(
  defaults,
  prompt = NULL,
  add_to_history = TRUE,
  prompt_build = TRUE,
  preview = FALSE,
  ...
)
}
\arguments{
\item{defaults}{Defaults object, generally puled from \code{tidychat_defaults()}}

\item{prompt}{The prompt to send to the LLM}

\item{add_to_history}{If supported by the LLM, indicates wether to include
this request in the history of request for this session}

\item{prompt_build}{Include the context and additional prompt as part of the
request}

\item{preview}{Primarily used for debugging. It indicates if it should send
the prompt to the LLM (FALSE), or if it should print out the resulting
prompt (TRUE)}

\item{...}{Optional arguments; currently unused.}
}
\description{
Method to easily integrate to new LLM's
}
\keyword{internal}
