% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.R
\name{tidychat_use_openai_gpt35}
\alias{tidychat_use_openai_gpt35}
\alias{tidychat_use_openai_davinci}
\title{Convenience functions to switch between models quickly}
\usage{
tidychat_use_openai_gpt35()

tidychat_use_openai_davinci()
}
\description{
Convenience functions to switch between models quickly
}
