% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychat.R
\name{tidychat_yaml_copy}
\alias{tidychat_yaml_copy}
\title{Creates}
\usage{
tidychat_yaml_copy(path = "config.yml", overwrite = FALSE)
}
\arguments{
\item{overwrite}{If there's an existing "config.yml", should it be replaced?}
}
\description{
Creates
}
