% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch-submit.R
\name{ch_submit}
\alias{ch_submit}
\title{Method to easily integrate to new LLM's}
\usage{
ch_submit(
  defaults,
  prompt = NULL,
  stream = NULL,
  prompt_build = TRUE,
  preview = FALSE,
  ...
)
}
\arguments{
\item{defaults}{Defaults object, generally puled from \code{chattr_defaults()}}

\item{prompt}{The prompt to send to the LLM}

\item{stream}{To output the response from the LLM as it happens, or wait until
the response is complete. Defaults to TRUE.}

\item{prompt_build}{Include the context and additional prompt as part of the
request}

\item{preview}{Primarily used for debugging. It indicates if it should send
the prompt to the LLM (FALSE), or if it should print out the resulting
prompt (TRUE)}

\item{...}{Optional arguments; currently unused.}
}
\description{
Method to easily integrate to new LLM's
}
\keyword{internal}
