% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chattr-test.R
\name{chattr_test}
\alias{chattr_test}
\title{Confirms conectivity to LLM interface}
\usage{
chattr_test(defaults = chattr_defaults())
}
\arguments{
\item{defaults}{Defaults object, generally puled from \code{chattr_defaults()}}
}
\description{
Confirms conectivity to LLM interface
}
