% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychat.R
\name{tidychat_prompt}
\alias{tidychat_prompt}
\alias{tidychat_debug_set_true}
\alias{tidychat_debug_set_false}
\alias{tidychat_debug_get}
\title{Debugging functions}
\usage{
tidychat_prompt(prompt = NULL)

tidychat_debug_set_true()

tidychat_debug_set_false()

tidychat_debug_get()
}
\arguments{
\item{prompt}{Request to send to LLM. Defaults to NULL}
}
\description{
Debugging functions
}
