% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-chat.R
\name{tidychat_prompt}
\alias{tidychat_prompt}
\title{Previews the prompt}
\usage{
tidychat_prompt(prompt = NULL)
}
\arguments{
\item{prompt}{Request to send to LLM. Defaults to NULL}
}
\description{
Previews the prompt
}
